refactor: Fastboot "support"
Previously the addon was not working on fastboot side, worse it makes the app crash. This was due to the fact that the import of bugsnag was also calling browser only variables which are not available on node side.

To fix this issue the addon now import bugsnag when needed and only on browser side. When called on nodejs, the addon will output an error message and redirect all logs to the console outpout